requirements.txt
azure-functions
Locales
Flask==1.1.2
requests==2.23.0